fix: coin balance history with internal tx changes
